Updated parseTypeEx() to return a type instead of token and dimensions.
The Type structure was extended to include the properties of an array.
Thus, users can access array attributes (and possibly functions in the future).

Also the new attribute identifier in the Type structure stores reference
to the token involved in the declaration of the type within the
compilation unit. This attribute is null for primitive types such as
arrays, integers, boolean, floats, strings, and arrays.

Added the boolean and the unsigned integer types to the Primitives structure.
A dummy type called the unknown was added, too.
Since the parser assigns type information to the context objects,
the unknown type is assigned in places where the type is not primitive.
The type is then updated in the analyzer.

Previously, context objects stored the base type and the dimensions seperately.
Now the context objects store type information using the Type structure,
which makes things somewhat simpler.