Updated the parser to follow variable declaration when a primitive type is used.
The isType() and followVariableDeclaration() did not consider the primitive types
as part of variable declaration. Therefore, the variableDeclaration rule was never
invoked previously. This commit fixes this issue. Further, the unsigned integer
and the boolean type were introduced to the isType() macro.

TODO: The string keyword should be added to isType() when it is implemented in
the lexer.