Updated the implementation of scopes.
The context objects are treated as symbols. You can define and resolve context
objects from scopes. To enable this a new structure called Symbol was introduced
in this commit, which consists of tag, name, and nameSize. These attributes allow
the scope to put and get context objects from the hash map.

Previously, the Variable structure was not a context. In order to make it operable
with the scope interface, it was converted to a scope in this commit.