The analyzer now reports undeclared types in variable declarations.
The parse does not have enough information to use the Type structure in variable
declarations. Therefore, a new structure called the VariableType was introduced.
It basially stores the identifier or keyword token in a type and the number
of dimensions in the case of arrays.
The analyzer is then responsible to convert the VariableType to Type during
resolution.

In this commit, the analyzer reports undeclared types in variable declarations
within structures. Other cases will soon be implemented.