The analyzer now reports undeclared types in function parameters.
Removed unnecessary prototypes and functions from the analyzer. Updated the protypes
for the local functions. Further, the order of the functions was changed to group
functions logically.

The invalidate() macro is a simple utility that invalidates the current scope
and sets it to the parent of the current scope. As the analyzer visits the AST
nodes, the current scope must be updated accordingly. Otherwise the resolution
and definition of symbols cannot be determined.