Implemented type inference
For variable declarations that expect type inference, the analyzer now resolves
the type of the specified initializer and infers the the variable type.
This allows users to declare different types of variables in the same statement.
Although I would not recommend this, but it is allowed.

Further, the analyzer now checks if the variable initializer is compatible with
the given variable type.