Analyze the return, throw, and break statements.
To check the compatibility of the return value with respect to the function,
we need access to the current function. Therefore, a new attribute `function`
was added to the analyzer. It keeps track of the current function we
are analyzing.

Added a few error messages and code to accomodate the errors that can be aused
when working with these statements.

The resolveThrowStatement() is incomplete. It needs to check if the expression
evaluates to an instance of the `Error` structure. Further, should we check
if the `code` attribute is set?

The compiler was updated to trigger the resolution phase only if there are no
errors in the definition phase.