Analyze type for array initializer expressions.
Array types are stored in the base type as a list. This allows us to
quickly create new array types.

Currently, we cannot infer types for empty array initializers. Therefore, it is reported as an error. Until this feature is implemented, users can
use the new expression or some allocation function from the standard library.