Implemented the Kush runtime.
The runtime includes a simple mark-sweep garbage collector and a custom allocator.
The allocator stores references to all free lists and allocated objects. All references
are stored in the `pointers` array in the current stackframe. During the marking phase,
the invocation stack is traversed marking these pointers. Any unmarked object from the
objects list is is deallocated during collection.