Updated the lexer to tokenize identifiers with colon.
This commit changes the lexer so that identifiers can now include one colon.
The rule for the new identifiers is two occurrences of a text recognized
by the previous identifier rule seperated by a colon, as shown below:
```
(PREVIOUS_IDENTIFIER ':' PREVIOUS_IDENTIFIER)
```

Only a single colon can appear in between. An error is generated, if the character
followed by the colon is not a letter.