Arrays can now be created with new expression and the array literal notation.
Primitive and reference arrays were added to the runtime. Right now,
both forms of arrays include an internal buffer, which requires an extra
allocation. This extra step can be removed with a little work. Further, only
i32 and reference arrays were implemented. The runtime needs a little refinement
to fix bugs and other improve the code quality.

The parser can now detect whether a given expression evaluates to either a
placeholder or a value. This information helps the compiler report an
error for invalid left-values in assignments.

The garbage collector switches the sense of the marking flag for each
consecutive collection. This way objects do not have to be unmarked
manually.

Finally, the generator now generates structure specific constructor prototypes.